mastodo_streamer/requirements.txt
flask~=3.0.3
pymongo
mastodon.py
python-dotenv